Move config firebase near to the services
